Initial implementation of cert-broker
Cert-Broker is a complementary component for Cert-Manager [1].
It enables certificate management for Kubernetes clusters which don't
operate their own (in-cluster) Cert-Manager, e.g. for organizational
resposibilities the Cert-Manager is located in another cluster.

[1] https://github.com/jetstack/cert-manager

This is the initial contribution to the Open Source Gardener project.

The following SAP developers contributed to this project until this
initial contribution was published as open source:

$ git shortlog -s -n
    41  Tim Usner
     3  Matthias Sohn

It is derived from the historical, SAP internal cert-broker repository
at commit ff98b41fe66809ca9ec59f420482c243c6013bba.